[wrapper] Move replay_channel to binius_spartan_prover
ReplayChannel and its WitnessGenWire CircuitWire backend are only used on
the prover side (by ZKWrappedProverChannel and crate prover's zk_config),
so the module belongs in spartan-prover, not spartan-verifier.

- crates/spartan-verifier/src/wrapper/replay_channel.rs →
  crates/spartan-prover/src/wrapper/replay_channel.rs (move tracked by jj
  via content similarity).
- replay_channel.rs's import for CircuitElem/CircuitWire changes from
  super::circuit_elem to binius_spartan_verifier::wrapper::circuit_elem.
  The IronSpartanBuilderChannel rustdoc cross-link is rewritten to point
  at the verifier crate path.
- spartan-prover's wrapper/mod.rs adds `pub mod replay_channel` and
  re-exports ReplayChannel; spartan-verifier's wrapper/mod.rs drops both.
- spartan-prover's wrapper/zk_wrapped_prover_channel.rs now imports
  ReplayChannel from `crate::wrapper`. The integration test in
  spartan-prover/tests/wrapper_integration_test.rs and the consumer in
  prover/src/zk_config.rs are updated to import ReplayChannel from
  binius_spartan_prover::wrapper rather than binius_spartan_verifier::
  wrapper.
- The verifier crate's circuit_elem module docs drop the rustdoc link to
  the now-absent WitnessGenWire/ReplayChannel types and reference them by
  path string instead (no cycle: prover depends on verifier, not vice
  versa).
- spartan-prover gains an `array-util` dependency (used inside the
  WitnessGenWire CircuitWire impl that moved with the module).
- The test_square_transpose_wires test moves from spartan-verifier's
  tests module into a new `#[cfg(test)] mod tests` inside the prover-side
  replay_channel.rs — it exercises both BuilderWire (verifier) and
  WitnessGenWire (prover), so it belongs in the prover crate (where both
  can be imported).

The `test_zk_wrapped_prove_verify` integration test in spartan-prover is
failing both on this commit and on its parent (036049a7); the panic is in
zk_wrapped_channel.rs's `verify_oracle_relations` precondition check and
is unrelated to this move.
