+++ crates/spartan-prover/src/wrapper/zk_wrapped_prover_channel.rs
@@ -23,7 +23,9 @@ use binius_iop_prover::{
 use binius_ip_prover::channel::IPProverChannel;
 use binius_math::{FieldBuffer, FieldSlice, ntt::AdditiveNTT};
 use binius_spartan_frontend::constraint_system::WitnessLayout;
-use binius_spartan_verifier::{IOPVerifier, wrapper::ReplayChannel};
+use binius_spartan_verifier::IOPVerifier;
+
+use crate::wrapper::ReplayChannel;
 use binius_transcript::fiat_shamir::Challenger;
 use binius_utils::SerializeBytes;
 use rand::CryptoRng;
